requirements.txt
alembic==1.9.2
blinker==1.5
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
python-dotenv==0.21.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
WTForms==3.0.1
